cAdvisor dependency + Dedicated Labels.
Made the following updates to the README.
1. Added sub-sub-section that talks about how cAdvisor prefixes all
container labels and that these prefixes should be included as part
of the labels provided for filtering.
2. Task Ranker is now compatible with cadvisor only.

Changed duration to unsigned int for Range.

Added dedicated label matchers that can be used filter prometheus data
based on a task's unique identifier and/or name of the host on which
a task is running.
Strategies can mandate the use of dedicated label matchers, especially
if tasks' runtime information is used for ranking.
LabelMatcher not contains a type that can be set to one of the following.
1. Other (default).
2. TaskID - flag label as one that can be used to fetch unique identifier
	of task.
3. TaskHostname - flag labe as one that can be used to fetch name of host
	on which task is running.

Updated README to include instructions on providing dedicated labels.

+++ query/builder_test.go
@@ -55,7 +55,7 @@ func TestGetBuilder(t *testing.T) {
 		Value:    "test_value2",
 	}, testBuilder.labelMatchers[1])
 	assert.Equal(t, Seconds, testBuilder.timeUnit)
-	assert.Equal(t, 5, testBuilder.timeDuration)
+	assert.Equal(t, uint(5), testBuilder.timeDuration)
 }
 
 func TestBuilder_BuildQuery(t *testing.T) {
